Rework rules and CSS
The rules weren't really the ones you'd get in the original game, so
this commit implements new states for the boxes, and one would lose the
game if they clicked on a box that's not ready to be deactivated.

Also, the CSS has been simplified and renders better on mobile now.